requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
websockets>=12.0
google-genai>=0.2.2
pyautogui>=0.9.54
pillow>=10.2.0
